% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distchange.R
\name{calc_distchange}
\alias{calc_distchange}
\title{Calculate distance change in future distribution}
\usage{
calc_distchange(
  eachfutureraster,
  raster.present.proj,
  eachspeciesname,
  crs.eqap,
  eachdirname
)
}
\arguments{
\item{presentraster}{historical raster}

\item{speciesname}{scientific name}

\item{dirname}{common species name}

\item{futurescenarioname1}{ssp126}

\item{futurescenarioname2}{ssp245}

\item{futurescenarioname3}{ssp585}
}
\value{
dist.change
}
\description{
Calculate distance change in future distribution
}
