% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-estadoscosteros.R
\docType{data}
\name{estadoscosteros}
\alias{estadoscosteros}
\title{DATASET_TITLE}
\format{
A data frame with 118886 rows and 236 variables:
\describe{
\item{\code{COM_ID}}{character COLUMN_DESCRIPTION}
\item{\code{ENTIDAD}}{double COLUMN_DESCRIPTION}
\item{\code{NOM_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{MUN}}{double COLUMN_DESCRIPTION}
\item{\code{NOM_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{LOC}}{double COLUMN_DESCRIPTION}
\item{\code{NOM_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{lon_deg}}{double COLUMN_DESCRIPTION}
\item{\code{lon_min}}{double COLUMN_DESCRIPTION}
\item{\code{lon_sec}}{double COLUMN_DESCRIPTION}
\item{\code{lat_deg}}{double COLUMN_DESCRIPTION}
\item{\code{lat_min}}{double COLUMN_DESCRIPTION}
\item{\code{lat_sec}}{double COLUMN_DESCRIPTION}
\item{\code{ALTITUD}}{double COLUMN_DESCRIPTION}
\item{\code{POBTOT}}{double COLUMN_DESCRIPTION}
\item{\code{POBFEM}}{double COLUMN_DESCRIPTION}
\item{\code{POBMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_0A2}}{double COLUMN_DESCRIPTION}
\item{\code{P_0A2_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_0A2_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_3YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_3YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_3YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_5YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_5YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_5YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_12YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_12YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_12YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_15YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_15YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_15YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_18YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_18YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_18YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_3A5}}{double COLUMN_DESCRIPTION}
\item{\code{P_3A5_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_3A5_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_6A11}}{double COLUMN_DESCRIPTION}
\item{\code{P_6A11_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_6A11_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_8A14}}{double COLUMN_DESCRIPTION}
\item{\code{P_8A14_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_8A14_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_12A14}}{double COLUMN_DESCRIPTION}
\item{\code{P_12A14_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_12A14_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_15A17}}{double COLUMN_DESCRIPTION}
\item{\code{P_15A17_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_15A17_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_18A24}}{double COLUMN_DESCRIPTION}
\item{\code{P_18A24_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_18A24_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_15A49}}{double COLUMN_DESCRIPTION}
\item{\code{P_60YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_60YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_60YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{REL_H_M}}{double COLUMN_DESCRIPTION}
\item{\code{POB0_14}}{double COLUMN_DESCRIPTION}
\item{\code{POB15_6}}{double COLUMN_DESCRIPTION}
\item{\code{POB65_M}}{double COLUMN_DESCRIPTION}
\item{\code{PROM_HN}}{double COLUMN_DESCRIPTION}
\item{\code{PNACENT}}{double COLUMN_DESCRIPTION}
\item{\code{PNACENT_F}}{double COLUMN_DESCRIPTION}
\item{\code{PNACENT_M}}{double COLUMN_DESCRIPTION}
\item{\code{PNACOE}}{double COLUMN_DESCRIPTION}
\item{\code{PNACOE_F}}{double COLUMN_DESCRIPTION}
\item{\code{PNACOE_M}}{double COLUMN_DESCRIPTION}
\item{\code{PRES2015}}{double COLUMN_DESCRIPTION}
\item{\code{PRES2015_F}}{double COLUMN_DESCRIPTION}
\item{\code{PRES2015_M}}{double COLUMN_DESCRIPTION}
\item{\code{PRESOE15}}{double COLUMN_DESCRIPTION}
\item{\code{PRESOE15_F}}{double COLUMN_DESCRIPTION}
\item{\code{PRESOE15_M}}{double COLUMN_DESCRIPTION}
\item{\code{P3YM_HLI}}{double COLUMN_DESCRIPTION}
\item{\code{P3YM_HLI_F}}{double COLUMN_DESCRIPTION}
\item{\code{P3YM_HLI_M}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLINHE}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLINHE_F}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLINHE_M}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLI_HE}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLI_HE_F}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLI_HE_M}}{double COLUMN_DESCRIPTION}
\item{\code{P5_HLI}}{double COLUMN_DESCRIPTION}
\item{\code{P5_HLI_N}}{double COLUMN_DESCRIPTION}
\item{\code{P5_HLI_H}}{double COLUMN_DESCRIPTION}
\item{\code{PHOG_IN}}{double COLUMN_DESCRIPTION}
\item{\code{POB_AFRO}}{double COLUMN_DESCRIPTION}
\item{\code{POB_AFRO_F}}{double COLUMN_DESCRIPTION}
\item{\code{POB_AFRO_M}}{double COLUMN_DESCRIPTION}
\item{\code{PCON_DI}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_MO}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_V}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_L}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_A}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_ME}}{double COLUMN_DESCRIPTION}
\item{\code{PCON_LI}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_C}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_V}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_H}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_O}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_M}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_R}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_P}}{double COLUMN_DESCRIPTION}
\item{\code{PSIND_L}}{double COLUMN_DESCRIPTION}
\item{\code{P3A5_NOA}}{double COLUMN_DESCRIPTION}
\item{\code{P3A5_NOA_F}}{double COLUMN_DESCRIPTION}
\item{\code{P3A5_NOA_M}}{double COLUMN_DESCRIPTION}
\item{\code{P6A11_NOA}}{double COLUMN_DESCRIPTION}
\item{\code{P6A11_NOAF}}{double COLUMN_DESCRIPTION}
\item{\code{P6A11_NOAM}}{double COLUMN_DESCRIPTION}
\item{\code{P12A14NOA}}{double COLUMN_DESCRIPTION}
\item{\code{P12A14NOAF}}{double COLUMN_DESCRIPTION}
\item{\code{P12A14NOAM}}{double COLUMN_DESCRIPTION}
\item{\code{P15A17A}}{double COLUMN_DESCRIPTION}
\item{\code{P15A17A_F}}{double COLUMN_DESCRIPTION}
\item{\code{P15A17A_M}}{double COLUMN_DESCRIPTION}
\item{\code{P18A24A}}{double COLUMN_DESCRIPTION}
\item{\code{P18A24A_F}}{double COLUMN_DESCRIPTION}
\item{\code{P18A24A_M}}{double COLUMN_DESCRIPTION}
\item{\code{P8A14AN}}{double COLUMN_DESCRIPTION}
\item{\code{P8A14AN_F}}{double COLUMN_DESCRIPTION}
\item{\code{P8A14AN_M}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_AN}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_AN_F}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_AN_M}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_SE}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_SE_F}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_SE_M}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_IN}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_INF}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_INM}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_CO}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_COF}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_COM}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_IN}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_INF}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_INM}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_CO}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_COF}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_COM}}{double COLUMN_DESCRIPTION}
\item{\code{P18YM_PB}}{double COLUMN_DESCRIPTION}
\item{\code{P18YM_PB_F}}{double COLUMN_DESCRIPTION}
\item{\code{P18YM_PB_M}}{double COLUMN_DESCRIPTION}
\item{\code{GRAPROES}}{double COLUMN_DESCRIPTION}
\item{\code{GRAPROES_F}}{double COLUMN_DESCRIPTION}
\item{\code{GRAPROES_M}}{double COLUMN_DESCRIPTION}
\item{\code{PEA}}{double COLUMN_DESCRIPTION}
\item{\code{PEA_F}}{double COLUMN_DESCRIPTION}
\item{\code{PEA_M}}{double COLUMN_DESCRIPTION}
\item{\code{PE_INAC}}{double COLUMN_DESCRIPTION}
\item{\code{PE_INAC_F}}{double COLUMN_DESCRIPTION}
\item{\code{PE_INAC_M}}{double COLUMN_DESCRIPTION}
\item{\code{POCUPADA}}{double COLUMN_DESCRIPTION}
\item{\code{POCUPADA_F}}{double COLUMN_DESCRIPTION}
\item{\code{POCUPADA_M}}{double COLUMN_DESCRIPTION}
\item{\code{PDESOCUP}}{double COLUMN_DESCRIPTION}
\item{\code{PDESOCUP_F}}{double COLUMN_DESCRIPTION}
\item{\code{PDESOCUP_M}}{double COLUMN_DESCRIPTION}
\item{\code{PSINDER}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_SS}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_IMSS}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_ISTE}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_ISTEE}}{double COLUMN_DESCRIPTION}
\item{\code{PAFIL_P}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_SE}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_IMSSB}}{double COLUMN_DESCRIPTION}
\item{\code{PAFIL_I}}{double COLUMN_DESCRIPTION}
\item{\code{PAFIL_O}}{double COLUMN_DESCRIPTION}
\item{\code{P12YM_SO}}{double COLUMN_DESCRIPTION}
\item{\code{P12YM_C}}{double COLUMN_DESCRIPTION}
\item{\code{P12YM_SE}}{double COLUMN_DESCRIPTION}
\item{\code{PCATOLI}}{double COLUMN_DESCRIPTION}
\item{\code{PRO_CRI}}{double COLUMN_DESCRIPTION}
\item{\code{POTRAS_}}{double COLUMN_DESCRIPTION}
\item{\code{PSIN_RE}}{double COLUMN_DESCRIPTION}
\item{\code{TOTHOG}}{double COLUMN_DESCRIPTION}
\item{\code{HOGJEF_F}}{double COLUMN_DESCRIPTION}
\item{\code{HOGJEF_M}}{double COLUMN_DESCRIPTION}
\item{\code{POBHOG}}{double COLUMN_DESCRIPTION}
\item{\code{PHOGJEF_F}}{double COLUMN_DESCRIPTION}
\item{\code{PHOGJEF_M}}{double COLUMN_DESCRIPTION}
\item{\code{VIVTOT}}{double COLUMN_DESCRIPTION}
\item{\code{TVIVHAB}}{double COLUMN_DESCRIPTION}
\item{\code{TVIVPAR}}{double COLUMN_DESCRIPTION}
\item{\code{VIVPAR_H}}{double COLUMN_DESCRIPTION}
\item{\code{TVIVPARH}}{double COLUMN_DESCRIPTION}
\item{\code{VIVPAR_D}}{double COLUMN_DESCRIPTION}
\item{\code{VIVPAR_U}}{double COLUMN_DESCRIPTION}
\item{\code{OCUPVIV}}{double COLUMN_DESCRIPTION}
\item{\code{PROM_OC}}{double COLUMN_DESCRIPTION}
\item{\code{PRO_OCU}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_PISOD}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_PISOT}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_1DO}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_2YM}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_1CU}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_2CU}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_3YM}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_C_E}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_S_E}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_AGUAD}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_AEA}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_AGUAF}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_TIN}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_CIS}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_EXC}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_LET}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_DRE}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_NOD}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_C_S}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_NDE}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_DSA}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_NDA}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SNB}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_REF}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_LAV}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_HMI}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_AUT}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_MOT}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_BIC}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_RAD}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_TV}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_PC}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_TEL}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_CEL}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_INT}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_STV}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SPM}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_CVJ}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SINR}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SINL}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SINC}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SINT}}{double COLUMN_DESCRIPTION}
\item{\code{TAMLOC}}{double COLUMN_DESCRIPTION}
\item{\code{id_row}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
estadoscosteros
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
