% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exposurehistorical.R
\docType{data}
\name{exposurehistorical}
\alias{exposurehistorical}
\title{DATASET_TITLE}
\format{
A data frame with 12905 rows and 26 variables:
\describe{
\item{\code{COM_ID}}{character COLUMN_DESCRIPTION}
\item{\code{dec_lon}}{double COLUMN_DESCRIPTION}
\item{\code{dec_lat}}{double COLUMN_DESCRIPTION}
\item{\code{impacto_antropogenico}}{double COLUMN_DESCRIPTION}
\item{\code{impacto_humano}}{double COLUMN_DESCRIPTION}
\item{\code{intervalo_diurno}}{double COLUMN_DESCRIPTION}
\item{\code{marea_roja}}{double COLUMN_DESCRIPTION}
\item{\code{oxigeno_disuelto}}{double COLUMN_DESCRIPTION}
\item{\code{precipitacion_promedio}}{double COLUMN_DESCRIPTION}
\item{\code{productividad_primaria}}{double COLUMN_DESCRIPTION}
\item{\code{riesgo_huracanes}}{double COLUMN_DESCRIPTION}
\item{\code{riesgo_inundaciones}}{double COLUMN_DESCRIPTION}
\item{\code{temperatura_promedio_anual}}{double COLUMN_DESCRIPTION}
\item{\code{temperatura_superficial_del_mar}}{double COLUMN_DESCRIPTION}
\item{\code{temperatura_superficial}}{double COLUMN_DESCRIPTION}
\item{\code{ENTIDAD}}{integer COLUMN_DESCRIPTION}
\item{\code{NOM_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{MUN}}{integer COLUMN_DESCRIPTION}
\item{\code{NOM_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{NOM_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{POBTOT}}{integer COLUMN_DESCRIPTION}
\item{\code{LOC}}{integer COLUMN_DESCRIPTION}
\item{\code{PC1}}{double COLUMN_DESCRIPTION}
\item{\code{PCA_norm}}{double COLUMN_DESCRIPTION}
\item{\code{region}}{character COLUMN_DESCRIPTION}
\item{\code{category}}{character COLUMN_DESCRIPTION}
}
}
\usage{
exposurehistorical
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
