% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lifehistory.R
\name{get_lifehistory}
\alias{get_lifehistory}
\title{Get lifehistory data}
\usage{
get_lifehistory(lifehistorydata, localities, fipinventory, lifehistorytraits)
}
\arguments{
\item{lifehistorydata}{life history characteristics}

\item{localities}{FIP locations}

\item{fipinventory}{Inventory of FIP projects from fisheryprogress.org}
}
\value{
lifehistory.trait
}
\description{
Get lifehistory data
}
