% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lifehistorytraits.R
\docType{data}
\name{lifehistorytraits}
\alias{lifehistorytraits}
\title{DATASET_TITLE}
\format{
A data frame with 31 rows and 11 variables:
\describe{
\item{\code{Type}}{character COLUMN_DESCRIPTION}
\item{\code{col_name}}{character COLUMN_DESCRIPTION}
\item{\code{quant_index}}{double COLUMN_DESCRIPTION}
\item{\code{ideal_directionality}}{character COLUMN_DESCRIPTION}
\item{\code{topsis_weight}}{double COLUMN_DESCRIPTION}
\item{\code{Trait}}{character COLUMN_DESCRIPTION}
\item{\code{Definition}}{character COLUMN_DESCRIPTION}
\item{\code{Low}}{list COLUMN_DESCRIPTION}
\item{\code{Moderate}}{list COLUMN_DESCRIPTION}
\item{\code{High}}{list COLUMN_DESCRIPTION}
\item{\code{Very high}}{list COLUMN_DESCRIPTION}
}
}
\usage{
lifehistorytraits
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
