% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_factor.R
\name{pca_factor}
\alias{pca_factor}
\title{Normalize function}
\usage{
pca_factor(thisdata, datasetname, namesdata)
}
\description{
Normalize from 0-1
}
\details{
INPUT: 1) data

OUTPUT: 1) normalized data
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna@gmail.com
}
