% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_googlesheet.R
\name{read_googlesheet}
\alias{read_googlesheet}
\title{Read life history Google sheet}
\usage{
read_googlesheet(sheetlink, sheetname, coltypes)
}
\arguments{
\item{sheetname}{name of sheet read}

\item{dataname}{name of data}
}
\value{
rds files for every sheet
}
\description{
Read life history Google sheet
}
\examples{
read_googlesheet(sheetlink, sheetname = "datos_caracteristicas")
}
