% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_indi.R
\name{scale_indi}
\alias{scale_indi}
\title{Scale variable}
\usage{
scale_indi(datatable, loc.fip.expo)
}
\arguments{
\item{sspscenario}{}
}
\value{
num.scale data with scaled total
}
\description{
Scale variable
}
