% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_var.R
\name{scale_var}
\alias{scale_var}
\title{Scale variable}
\usage{
scale_var(datatable, sspscenario)
}
\arguments{
\item{sspscenario}{}
}
\value{
num.scale data with scaled total
}
\description{
Scale variable
}
\examples{
scaled.exp <- scale_var(exp.data.set.sc, NA)
}
