% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_var_tab.R
\name{select_var_tab}
\alias{select_var_tab}
\title{Select variables}
\usage{
select_var_tab(namesdata, datatable, scenario, id.cols, expohist.bc)
}
\arguments{
\item{sspscenario}{}
}
\value{
num.scale data with scaled total
}
\description{
Select variables
}
