% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fipinventory.R
\docType{data}
\name{fipinventory}
\alias{fipinventory}
\title{DATASET_TITLE}
\format{
A data frame with 50 rows and 14 variables:
\describe{
\item{\code{fip_name}}{character COLUMN_DESCRIPTION}
\item{\code{region}}{character COLUMN_DESCRIPTION}
\item{\code{species}}{character COLUMN_DESCRIPTION}
\item{\code{common_name_en}}{character COLUMN_DESCRIPTION}
\item{\code{common_name}}{character COLUMN_DESCRIPTION}
\item{\code{project_start_end_dates}}{character COLUMN_DESCRIPTION}
\item{\code{estimated_total_fip_landings_mt}}{double COLUMN_DESCRIPTION}
\item{\code{estimated_total_fishery_landings_mt}}{double COLUMN_DESCRIPTION}
\item{\code{landings_date}}{double COLUMN_DESCRIPTION}
\item{\code{organization_name}}{character COLUMN_DESCRIPTION}
\item{\code{fip_id_number}}{integer COLUMN_DESCRIPTION}
\item{\code{includes_climate_change}}{character COLUMN_DESCRIPTION}
\item{\code{climate_change_requirement}}{character COLUMN_DESCRIPTION}
\item{\code{proponent}}{character COLUMN_DESCRIPTION}
}
}
\usage{
fipinventory
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
